% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4ds.tutorials-package.R
\docType{package}
\name{r4ds.tutorials-package}
\alias{r4ds.tutorials}
\alias{r4ds.tutorials-package}
\title{r4ds.tutorials: Tutorials for R for Data Science}
\description{
When assigned "R for Data Science" (Wickham, Çetinkaya-Rundel, and Grolemund (2023, ISBN: 1492097402)), students should read the book and type in all the associated R commands themselves. Sadly, that never happens. These tutorials allow students to demonstrate (and their instructors to be sure) that all work has been completed. See Kane (2023) \url{https://ppbds.github.io/tutorial.helpers/articles/instructions.html} from the 'tutorial.helpers' package for a background discussion.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ppbds.github.io/r4ds.tutorials/}
  \item \url{https://github.com/PPBDS/r4ds.tutorials}
  \item Report bugs at \url{https://github.com/PPBDS/r4ds.tutorials/issues}
}

}
\author{
\strong{Maintainer}: David Kane \email{dave.kane@gmail.com} (\href{https://orcid.org/0000-0002-6660-3934}{ORCID}) [copyright holder]

}
\keyword{internal}
