% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4ds.tutorials-package.R
\docType{package}
\name{r4ds.tutorials-package}
\alias{r4ds.tutorials}
\alias{r4ds.tutorials-package}
\title{r4ds.tutorials: Tutorials for R for Data Science}
\description{
Tutorials for R for Data Science by Hadley Wickham, Mine Çetinkaya-Rundel, and Garrett Grolemund. In an ideal world, students would read the book and type in all the associated R commands themselves. Sadly, that often does not happen. These tutorials allow students to demonstrate (and their instructors to be sure) that all work has been completed. See the tutorial.helpers package for a background discussion of the tool and approach.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ppbds.github.io/r4ds.tutorials/}
  \item Report bugs at \url{https://github.com/ppbds/r4ds.tutorials/issues}
}

}
\author{
\strong{Maintainer}: David Kane \email{dave.kane@gmail.com} (\href{https://orcid.org/0000-0002-6660-3934}{ORCID})

}
\keyword{internal}
